refactor(runtime): extract scriptable engine + impl from context_engine
The `ScriptableContextEngine` cluster moves out of `context_engine.rs`
into `context_engine/scriptable/`:

- `scriptable/mod.rs` (1,922 LOC) — struct definition, `impl
  ScriptableContextEngine` block (constructor + helper methods incl.
  the `pub(super) run_hook` exercised by tests), hook telemetry
  surface (`HookTrace`, `HookStats`, `HookMetrics`,
  `TRACE_BUFFER_CAPACITY`), per-hook circuit breaker
  (`CircuitBreakerState`) + rate limiter (`HookRateLimiter`), plugin
  loader (`plugins_dir`, `validate_plugin_name`, `load_plugin`).
- `scriptable/engine.rs` (1,064 LOC) — the `impl ContextEngine for
  ScriptableContextEngine` block with all eight hook methods
  (`bootstrap`, `ingest`, `assemble`, `compact`, `after_turn`,
  `prepare_subagent_context`, `merge_subagent_context`,
  `truncate_tool_result`). Split into its own file because the eight
  methods together come in at ~1k LOC; keeping them next to the
  struct would push the parent file over the 2k cap.

context_engine.rs:
  - `pub use self::scriptable::{plugins_dir, HookMetrics, HookStats,
    HookTrace, ScriptableContextEngine}` re-exports the public
    surface so existing call sites (kernel context-engine API,
    `trace_store::HookTrace` import) keep resolving the same paths.
  - `use self::scriptable::{load_plugin, TRACE_BUFFER_CAPACITY}` for
    the factory / metrics path still in the parent file.
  - `run_hook` is `pub(super)` so `context_engine/tests.rs` keeps
    invoking `ScriptableContextEngine::run_hook(...)` directly.

context_engine.rs: 5,218 -> 1,651 LOC (-3,567). Now under the 2,000
LOC cap from #3710's acceptance criteria.

cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
